requirements.txt
kwargs==1.0.1
mysql_connector_repackaged==0.3.1
pytest==6.2.3
requests==2.26.0
selenium==4.4.3
starlette==0.16.0
